% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.conformal.rho.R
\name{my.conformal.rho}
\alias{my.conformal.rho}
\title{Transform shape with curvature information}
\usage{
my.conformal.rho(vertices, faces.v, rho.v, face = FALSE)
}
\arguments{
\item{vertices}{is a quaternion vector, coordinates of vertices}

\item{faces.v}{is a 3 x No.faces integer matrix}

\item{rho.v}{is a scaler vector of length = no. vertices, or no. faces}

\item{face}{is a logical indicater rho.v is vertices vs. faces}
}
\description{
transfromation with curvature
}
\keyword{curvature}
\keyword{transformation}

