% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.make.E.v.R
\name{my.make.E.v}
\alias{my.make.E.v}
\title{Make E-matrix}
\usage{
my.make.E.v(vertices, faces.v, rho)
}
\description{
Four Real-valued matrices representing four components of quatenion
Returned value is a list of 4 sparse vectors
}

