% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.sphere.tri.mesh.R
\name{my.sphere.tri.mesh}
\alias{my.sphere.tri.mesh}
\title{Triangular mesh of sphere}
\usage{
my.sphere.tri.mesh(n.psi = 30)
}
\arguments{
\item{n.psi, }{integer to set fineness of mesh}
}
\description{
Triangular mesh of sphere
}
\examples{
n.psi <- 25
sp.mesh <- my.sphere.tri.mesh(n.psi)
plot3d(sp.mesh$xyz)
segments3d(sp.mesh$xyz[c(t(sp.mesh$edge)),])
}
\keyword{mesh}
\keyword{sphere}
\keyword{triangle}

