% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.sphereConformal.R
\name{my.sphereConformal}
\alias{my.sphereConformal}
\title{Conformal change from sphere}
\usage{
my.sphereConformal(n.psi, rho.fx)
}
\arguments{
\item{n.psi}{is fineness index of triangle mesh of sphere}

\item{rho.fx}{is a function giving rho value for coordinates on sphere}
}
\description{
conformal change from sphere
}
\examples{
n.psi <- 30 # ‹…–ÊƒƒbƒVƒ…‚ÌˆÜ“x‚Ý”
rho.fx <- function(X){
  ret <- sin(X[,3]*pi*2 )*3
  return(ret)
}
out1 <- my.sphereConformal(n.psi,rho.fx)
plot.sp.conformal(out1$xyz.ori,out1$faces.v,out1$sp.mesh$n.triangles,out1$rho.f)
}
\keyword{curvature}
\keyword{transformation}

