% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.vector.access.R
\name{my.vector.access}
\alias{my.vector.access}
\title{Utility function to apply functions to elements in sparse matrix}
\usage{
my.vector.access(v, a, func = sum, zero = 0)
}
\description{
utility function
}

