% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.fromVtoTri.R
\name{rho.fromVtoTri}
\alias{rho.fromVtoTri}
\title{Calculate rho of triangle from rho values of veritices}
\usage{
rho.fromVtoTri(rho.v, faces.v)
}
\arguments{
\item{rho.v}{is a scaler vector of rho of vertices}

\item{faces.v}{is a 3 x No.faces integer matrix}
}
\description{
Rho of triangle
}
\keyword{rho}

